Added basic route matching and rate limiting.
