+++ gpt-summary/package.json
@@ -4,7 +4,7 @@
   "description": "Chart with text summary for screen reader accessibility",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon index.js"
+    "start": "nodemon server.js"
   },
   "keywords": [
     "chatGPT",
@@ -15,7 +15,8 @@
   "author": "Trevor Pierce",
   "license": "MIT",
   "dependencies": {
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "express-rate-limit": "^6.7.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.22"
