+++ gpt-summary/public/404.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>404: Page not found</title>
+  <link rel="stylesheet" href="css/main.css">
+</head>
+
+<body>
+  <h1>Aw, frowny face</h1>
+  <p>This is a 404. The page you&rsquo;re looking for is not availalbe. Maybe try the <a href="/">front page</a>
+    instead?</p>
+</body>
+
+</html>
